+++ src/extension.ts
@@ -3,6 +3,7 @@ import { ICommandable } from "./interfaces/commandable";
 import { PrettifyDocumentCommand } from "./commands/prettifyDocumentCommand";
 import { MinifyDocumentCommand } from "./commands/minifyDocumentCommand";
 import { ToggleInlayHintsCommand } from "./commands/toggleInlayHintsCommand";
+import { SelectTextByPositionCommand } from "./commands/selectTextByPositionCommand";
 import { IProvidable } from "./interfaces/providable";
 import { HighlightEdiProvider } from "./providers/highlightEdiProvider";
 import { HoverX12Provider } from "./providers/hoverX12Provider";
@@ -18,6 +19,7 @@ export function activate(context: vscode.ExtensionContext) {
   registerCommand(context, new PrettifyDocumentCommand());
   registerCommand(context, new MinifyDocumentCommand());
   registerCommand(context, new ToggleInlayHintsCommand());
+  registerCommand(context, new SelectTextByPositionCommand());
   registerProvider(context, new HighlightEdiProvider());
   registerProvider(context, new HoverX12Provider());
   registerProvider(context, new HoverEdifactProvider());